+++ src/test/java/chess/domain/board/BoardTest.java
@@ -11,7 +11,6 @@ import org.junit.jupiter.api.Test;
 
 import java.util.List;
 
-import static chess.domain.game.Team.BLACK;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 
 public class BoardTest {
@@ -44,52 +43,5 @@ public class BoardTest {
         ChessPiece targetChessPiece = board.getBoard().get(4 - CORRECTION_VALUE).get(1 - CORRECTION_VALUE);
         assertThat(startChessPiece).isInstanceOf(Blank.class);
         assertThat(targetChessPiece).isInstanceOf(Pawn.class);
-    }
-
-    @Test
-    @DisplayName("isGameEnd 테스트 - 게임이 끝났을 경우")
-    void isGameEnd() {
-        Board board = BoardFactory.createBoard();
-
-        board.move(new MovingInfo(Position.of(2, 6), Position.of(3, 6)));
-        board.move(new MovingInfo(Position.of(7, 5), Position.of(6, 5)));
-        board.move(new MovingInfo(Position.of(1, 5), Position.of(2, 6)));
-        board.move(new MovingInfo(Position.of(8, 4), Position.of(4, 8)));
-        board.move(new MovingInfo(Position.of(2, 6), Position.of(3, 7)));
-        board.move(new MovingInfo(Position.of(4, 8), Position.of(3, 7)));
-        assertThat(board.isGameEnd()).isTrue();
-    }
-
-    @Test
-    @DisplayName("isGameEnd 테스트 - 게임이 끝나지 않았을 경우")
-    void isGameEnd_IfNotEnd_ReturnFalse() {
-        Board board = BoardFactory.createBoard();
-
-        board.move(new MovingInfo(Position.of(2, 6), Position.of(3, 6)));
-        board.move(new MovingInfo(Position.of(7, 5), Position.of(6, 5)));
-        board.move(new MovingInfo(Position.of(1, 5), Position.of(2, 6)));
-        board.move(new MovingInfo(Position.of(8, 4), Position.of(4, 8)));
-        board.move(new MovingInfo(Position.of(2, 6), Position.of(3, 7)));
-        assertThat(board.isGameEnd()).isFalse();
-    }
-
-    @Test
-    @DisplayName("getTotalScore 테스트")
-    void getTotalScore() {
-        Board board = BoardFactory.createBoard();
-
-        board.move(new MovingInfo(Position.of(2, 4), Position.of(4, 4)));
-        board.move(new MovingInfo(Position.of(7, 5), Position.of(5, 5)));
-        board.move(new MovingInfo(Position.of(4, 4), Position.of(5, 5)));
-        assertThat(board.getTotalScore()).isEqualTo(37);
-    }
-
-    @Test
-    @DisplayName("getNowPlayingTeam 테스트")
-    void getNowPlayingTeam() {
-        Board board = BoardFactory.createBoard();
-
-        board.move(new MovingInfo(Position.of(2, 4), Position.of(4, 4)));
-        assertThat(board.getNowPlayingTeam()).isEqualTo(BLACK);
     }
 }
